refactor: member, auth 패키지 분리
