+++ src/main/java/nextstep/auth/application/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package nextstep.member.application;
+package nextstep.auth.application;
 
 import io.jsonwebtoken.*;
 import org.springframework.beans.factory.annotation.Value;
