+++ src/main/java/nextstep/auth/application/dto/GithubLoginRequest.java
@@ -1,4 +1,4 @@
-package nextstep.member.application.dto;
+package nextstep.auth.application.dto;
 
 
 public class GithubLoginRequest {
