+++ src/test/java/nextstep/auth/AuthAcceptanceTest.java
@@ -1,4 +1,4 @@
-package nextstep.member.acceptance;
+package nextstep.auth;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
@@ -10,11 +10,11 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 
-import static nextstep.member.acceptance.AuthSteps.깃헙_로그인_토큰_생성;
-import static nextstep.member.acceptance.AuthSteps.토큰_생성;
-import static nextstep.member.acceptance.GithubResponse.사용자1;
-import static nextstep.member.acceptance.MemberSteps.내_정보_요청;
-import static nextstep.member.acceptance.MemberSteps.토큰_없이_내_정보_요청;
+import static nextstep.auth.AuthSteps.깃헙_로그인_토큰_생성;
+import static nextstep.auth.AuthSteps.토큰_생성;
+import static nextstep.auth.GithubResponse.사용자1;
+import static nextstep.member.MemberSteps.내_정보_요청;
+import static nextstep.member.MemberSteps.토큰_없이_내_정보_요청;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class AuthAcceptanceTest extends AcceptanceTest {
