+++ src/test/java/nextstep/auth/GithubClientTest.java
@@ -1,14 +1,14 @@
-package nextstep.member.acceptance;
+package nextstep.auth;
 
-import nextstep.member.application.GithubClient;
-import nextstep.member.application.dto.GithubProfileResponse;
+import nextstep.auth.application.GithubClient;
+import nextstep.auth.application.dto.GithubProfileResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 
-import static nextstep.member.acceptance.GithubResponse.사용자1;
+import static nextstep.auth.GithubResponse.사용자1;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @ActiveProfiles("test")
