+++ src/test/java/nextstep/member/MemberAcceptanceTest.java
@@ -1,4 +1,4 @@
-package nextstep.member.acceptance;
+package nextstep.member;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
@@ -7,8 +7,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
-import static nextstep.member.acceptance.AuthSteps.토큰_생성;
-import static nextstep.member.acceptance.MemberSteps.*;
+import static nextstep.auth.AuthSteps.토큰_생성;
+import static nextstep.member.MemberSteps.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class MemberAcceptanceTest extends AcceptanceTest {
